+++ middleware.go
@@ -291,8 +291,8 @@ func (s *Service[SiteT]) parseForm(queryKey, rawQueryKey string) func(next http.
 
 			if len(queryForm) > 0 {
 				var qs string
-				if s.router.EncodeQuery != nil {
-					qs = s.router.EncodeQuery(queryForm)
+				if s.router.EncodeQuery() != nil {
+					qs = s.router.EncodeQuery()(queryForm)
 				} else {
 					qs = queryForm.Encode()
 				}